wip: airdrop store actions
